The "annoyingservice" package was renamed to "annoynacehandlers"
